modify order handling logic added order service
